// app.client/src/api/models/GetInventoryItemDetailByCodeQueryResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { InventoryItemDetailByCodeVM } from './InventoryItemDetailByCodeVM';
export type GetInventoryItemDetailByCodeQueryResponse = {
    success?: boolean;
    message?: string | null;
    validationErrors?: Array<string> | null;
    data?: InventoryItemDetailByCodeVM;
};